+++ lib/Elastica/QueryBuilder/DSL/Query.php
@@ -187,6 +187,8 @@ class Query implements DSL
     /**
      * filtered query.
      *
+     * @deprecated Use bool() instead. Filtered query is deprecated since ES 2.0.0-beta1 and this method will be removed in further Elastica releases.
+     *
      * @param AbstractFilter $filter
      * @param AbstractQuery  $query
      *
@@ -196,6 +198,7 @@ class Query implements DSL
      */
     public function filtered(AbstractQuery $query = null, AbstractFilter $filter = null)
     {
+        trigger_error('Use bool() instead. Filtered query is deprecated since ES 2.0.0-beta1 and this method will be removed in further Elastica releases.', E_USER_DEPRECATED);
         return new Filtered($query, $filter);
     }