+++ test/lib/Elastica/Test/Query/BoolQueryTest.php
@@ -2,10 +2,11 @@
 namespace Elastica\Test\Query;
 
 use Elastica\Document;
+use Elastica\Filter\Term as TermFilter;
 use Elastica\Index;
 use Elastica\Query\BoolQuery;
 use Elastica\Query\Ids;
-use Elastica\Query\Term;
+use Elastica\Query\Term as TermQuery;
 use Elastica\Test\Base as BaseTest;
 use Elastica\Type;
 
@@ -27,6 +28,12 @@ class BoolQueryTest extends BaseTest
         $idsQuery3 = new Ids();
         $idsQuery3->setIds(3);
 
+        $filter1 = new TermFilter();
+        $filter1->setTerm('test', '1');
+
+        $filter2 = new TermFilter();
+        $filter2->setTerm('username', 'ruth');
+
         $boost = 1.2;
         $minMatch = 2;
 
@@ -35,11 +42,14 @@ class BoolQueryTest extends BaseTest
         $query->addMust($idsQuery1);
         $query->addMustNot($idsQuery2);
         $query->addShould($idsQuery3->toArray());
+        $query->addFilter($filter1);
+        $query->addFilter($filter2);
 
         $expectedArray = array(
             'bool' => array(
                 'must' => array($idsQuery1->toArray()),
                 'should' => array($idsQuery3->toArray()),
+                'filter' => array($filter1->toArray(), $filter2->toArray()),
                 'minimum_number_should_match' => $minMatch,
                 'must_not' => array($idsQuery2->toArray()),
                 'boost' => $boost,
@@ -60,10 +70,10 @@ class BoolQueryTest extends BaseTest
     {
         $boolQuery = new BoolQuery();
 
-        $term1 = new Term();
+        $term1 = new TermQuery();
         $term1->setParam('interests', 84);
 
-        $term2 = new Term();
+        $term2 = new TermQuery();
         $term2->setParam('interests', 92);
 
         $boolQuery->addShould($term1)->addShould($term2);
@@ -83,36 +93,44 @@ class BoolQueryTest extends BaseTest
 
         $type = new Type($index, 'helloworld');
 
-        $doc = new Document(1, array('id' => 1, 'email' => 'hans@test.com', 'username' => 'hans', 'test' => array('2', '3', '5')));
+        $doc = new Document(1, array('id' => 1, 'email' => 'hans@test.com', 'username' => 'hans', 'test' => array('2', '4', '5')));
         $type->addDocument($doc);
         $doc = new Document(2, array('id' => 2, 'email' => 'emil@test.com', 'username' => 'emil', 'test' => array('1', '3', '6')));
         $type->addDocument($doc);
         $doc = new Document(3, array('id' => 3, 'email' => 'ruth@test.com', 'username' => 'ruth', 'test' => array('2', '3', '7')));
+        $type->addDocument($doc);
+        $doc = new Document(4, array('id' => 4, 'email' => 'john@test.com', 'username' => 'john', 'test' => array('2', '4', '8')));
         $type->addDocument($doc);
 
         // Refresh index
         $index->refresh();
 
         $boolQuery = new BoolQuery();
-        $termQuery1 = new Term(array('test' => '2'));
+        $termQuery1 = new TermQuery(array('test' => '2'));
         $boolQuery->addMust($termQuery1);
+        $resultSet = $type->search($boolQuery);
+
+        $this->assertEquals(3, $resultSet->count());
+
+        $termFilter = new TermFilter(array('test' => '4'));
+        $boolQuery->addFilter($termFilter);
         $resultSet = $type->search($boolQuery);
 
         $this->assertEquals(2, $resultSet->count());
 
-        $termQuery2 = new Term(array('test' => '5'));
+        $termQuery2 = new TermQuery(array('test' => '5'));
         $boolQuery->addMust($termQuery2);
         $resultSet = $type->search($boolQuery);
 
         $this->assertEquals(1, $resultSet->count());
 
-        $termQuery3 = new Term(array('username' => 'hans'));
+        $termQuery3 = new TermQuery(array('username' => 'hans'));
         $boolQuery->addMust($termQuery3);
         $resultSet = $type->search($boolQuery);
 
         $this->assertEquals(1, $resultSet->count());
 
-        $termQuery4 = new Term(array('username' => 'emil'));
+        $termQuery4 = new TermQuery(array('username' => 'emil'));
         $boolQuery->addMust($termQuery4);
         $resultSet = $type->search($boolQuery);